backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.4
chardet==4.0.0
cloudinary==1.25.0
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docopt==0.6.2
drf-jwt==1.17.3
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.9
pbr==5.5.1
Pillow==8.1.0
pip-upgrader==1.4.15
pipenv==2020.11.15
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
stevedore==3.3.0
tablib==3.0.0
terminaltables==3.1.0
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.3.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
